[修改] 完善了鉴权功能，写了个以个测试接口。里面的testPermission接口是有效权限，testPermissionfff是无效权限。 有效权限的可以正常访问，无效权限的无法访问。之后考虑给图片的端口设上权限注解。

+++ src/main/java/com/cl/fun/gag/pic/dao/UserMapper.java
@@ -16,6 +16,6 @@ public interface UserMapper {
 
     int insertUser(UserPo userPo);
 
-    List<UmsPermission> getUserPermission(@Param("userId") Integer userId);
+    List<UmsPermission> getUserPermission(@Param("userId") Long userId);
 
 }
